requirements.txt
# Core Dependencies
streamlit>=1.28.0
psycopg2-binary>=2.9.9
groq>=0.4.0
numpy>=1.24.3
pandas>=2.0.3

# Machine Learning & NLP
sentence-transformers>=2.2.2
scikit-learn>=1.3.0
nltk>=3.8.1
rank-bm25>=0.2.2
rouge-score>=0.1.2

# Vector Database
pgvector>=0.2.0

# Evaluation & Metrics
tqdm>=4.66.1
evaluate>=0.4.0

# Async Support
aiohttp>=3.8.5
asyncio>=3.4.3

# Logging & Monitoring
python-json-logger>=2.0.7

# Development & Testing
pytest>=7.4.0
black>=23.7.0
flake8>=6.1.0
mypy>=1.5.1
pytest-asyncio>=0.21.1

# Type Checking
typing-extensions>=4.7.1

# Documentation
sphinx>=7.1.2
sphinx-rtd-theme>=1.3.0

# Environment & Configuration
python-dotenv>=1.0.0
pyyaml>=6.0.1